backend/pyproject.toml
[project]
name = "techflow-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "feedparser>=6.0.11",
    "anthropic>=0.25.0",
    "openai>=1.30.0",
    "sqlmodel>=0.0.19",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
]

[tool.uv]
package = false
